MSD: Port Sidebar component into responsive-sidebar
Move the full Navigator-based Sidebar (with route sync, error boundary,
and inline PrimaryMenu) into responsive-sidebar/sidebar.tsx. Move
navigator-route-sync and error files into the responsive-sidebar folder.
