+++ client/dashboard/app/responsive-sidebar/navigator-route-sync.tsx
