+++ client/dashboard/app/responsive-sidebar/sidebar.tsx
@@ -1,20 +1,60 @@
+import { useRouterState } from '@tanstack/react-router';
+import { __experimentalHStack as HStack, Navigator } from '@wordpress/components';
 import { __, sprintf } from '@wordpress/i18n';
+import { brush, copy, envelope, globe, plugins } from '@wordpress/icons';
+import { useEffect, useMemo, useRef } from 'react';
 import RouterLinkButton from '../../components/router-link-button';
+import { SidebarExpandableMenuItem, SidebarMenu, SidebarMenuItem } from '../../components/sidebar';
+import DomainSidebar from '../../domains/domain-sidebar';
+import MeSidebar from '../../me/me-sidebar';
+import SiteSidebar from '../../sites/site-sidebar';
+import { wpcomLink } from '../../utils/link';
 import { useAnalytics } from '../analytics';
 import { useAppContext } from '../context';
+import RouteErrorBoundary from './error';
+import { getScreenPath, NavigatorRouteSync } from './navigator-route-sync';
 
 import './sidebar.scss';
 
-export default function Sidebar() {
+export default function Sidebar( { onNavigate }: { onNavigate?: () => void } ) {
 	const { Logo, name } = useAppContext();
 	const { recordTracksEvent } = useAnalytics();
+	const { resolvedPathname, hasError } = useRouterState( {
+		select: ( state ) => ( {
+			resolvedPathname: state.resolvedLocation?.pathname ?? state.location.pathname,
+			hasError: state.matches.some(
+				( match ) => match.status === 'error' || match.status === 'notFound'
+			),
+		} ),
+	} );
+	const screenPath = getScreenPath( resolvedPathname, hasError );
+	const initialPath = useMemo( () => screenPath, [] ); // eslint-disable-line react-hooks/exhaustive-deps
+
+	// Close responsive panel when a navigation link is clicked
+	const sidebarRef = useRef< HTMLDivElement >( null );
+	useEffect( () => {
+		if ( ! onNavigate ) {
+			return;
+		}
+		const el = sidebarRef.current;
+		if ( ! el ) {
+			return;
+		}
+		const handler = ( event: MouseEvent ) => {
+			if ( ( event.target as HTMLElement ).closest( 'a' ) ) {
+				onNavigate();
+			}
+		};
+		el.addEventListener( 'click', handler );
+		return () => el.removeEventListener( 'click', handler );
+	}, [ onNavigate ] );
 
 	return (
-		<div className="dashboard-responsive-sidebar__sidebar">
+		<div className="dashboard-responsive-sidebar__sidebar" ref={ sidebarRef }>
 			{ Logo && (
 				<div className="dashboard-responsive-sidebar__logo">
 					<RouterLinkButton
-						/* translators: Screen reader text for link to root of the hosting dashboard. "name" is the product of whose hosting dashboard this is: e.g. WordPress.com */
+						/* translators: Screen reader text for link to root of the hosting dashboard. "name" is the product name, e.g. WordPress.com */
 						aria-label={ sprintf( __( '%(name)s home' ), { name } ) }
 						icon={ <Logo /> }
 						to="/"
@@ -24,6 +64,86 @@ export default function Sidebar() {
 					/>
 				</div>
 			) }
+			<Navigator initialPath={ initialPath }>
+				<NavigatorRouteSync screenPath={ screenPath } />
+
+				<Navigator.Screen path="/">
+					<PrimaryMenuSidebar />
+				</Navigator.Screen>
+
+				<Navigator.Screen path="/sites/:siteSlug">
+					<RouteErrorBoundary>
+						<SiteSidebar />
+					</RouteErrorBoundary>
+				</Navigator.Screen>
+
+				<Navigator.Screen path="/domains/:domainName">
+					<RouteErrorBoundary>
+						<DomainSidebar />
+					</RouteErrorBoundary>
+				</Navigator.Screen>
+
+				<Navigator.Screen path="/me">
+					<RouteErrorBoundary>
+						<MeSidebar />
+					</RouteErrorBoundary>
+				</Navigator.Screen>
+			</Navigator>
 		</div>
+	);
+}
+
+function PrimaryMenuSidebar() {
+	const { supports } = useAppContext();
+
+	return (
+		<SidebarMenu>
+			{ supports.sites && (
+				<SidebarMenuItem icon={ copy } to="/sites">
+					{ __( 'Sites' ) }
+				</SidebarMenuItem>
+			) }
+			{ supports.domains && (
+				<SidebarMenuItem icon={ globe } to="/domains">
+					{ __( 'Domains' ) }
+				</SidebarMenuItem>
+			) }
+			{ supports.emails && (
+				<SidebarMenuItem icon={ envelope } to="/emails">
+					{ __( 'Emails' ) }
+				</SidebarMenuItem>
+			) }
+			{ supports.plugins && (
+				<SidebarExpandableMenuItem label={ __( 'Plugins' ) } icon={ plugins } to="/plugins">
+					<SidebarMenuItem to="/plugins/manage">{ __( 'Manage plugins' ) }</SidebarMenuItem>
+					<SidebarMenuItem to="/plugins/scheduled-updates">
+						{ __( 'Scheduled updates' ) }
+					</SidebarMenuItem>
+					<SidebarMenuItem
+						href={ wpcomLink( '/plugins' ) }
+						target="_blank"
+						rel="noopener noreferrer"
+					>
+						<HStack justify="flex-start" spacing={ 1 }>
+							<span>{ __( 'Browse plugins' ) }</span>
+							<span aria-label={ __( '(opens in a new tab)' ) }>&#8599;</span>
+						</HStack>
+					</SidebarMenuItem>
+				</SidebarExpandableMenuItem>
+			) }
+			{ supports.themes && (
+				<SidebarMenuItem
+					icon={ brush }
+					href={ wpcomLink( '/themes' ) }
+					target="_blank"
+					rel="noopener noreferrer"
+				>
+					<HStack justify="flex-start" spacing={ 1 }>
+						<span>{ __( 'Themes' ) }</span>
+						<span aria-label={ __( '(opens in a new tab)' ) }>&#8599;</span>
+					</HStack>
+				</SidebarMenuItem>
+			) }
+		</SidebarMenu>
 	);
 }
